Restructure packages/server directory
