+++ packages/server/src/mediaStreamsInfo.ts
@@ -1,0 +1,22 @@
+
+export interface MediaStreamsInfoSsrc {
+  ssrc: string;
+  cname?: string;
+  mslabel?: string;
+  label?: string;
+}
+
+export interface MediaStreamsInfoSsrcGroup {
+  semantics: string;
+  ssrcs: string[];
+}
+
+export interface MediaStreamsInfo {
+  audio: {
+    ssrcs: MediaStreamsInfoSsrc[];
+  },
+  video: {
+    ssrcs: MediaStreamsInfoSsrc[];
+    ssrcGroups: MediaStreamsInfoSsrcGroup[];
+  }
+}
