--- packages/server/src/models/WHIPResource.ts
@@ -1,45 +1,0 @@
-import { Broadcaster } from "../broadcaster";
-
-export const IANA_PREFIX = "urn:ietf:params:whip:";
-
-export interface WHIPResourceICEServer {
-    urls: string;
-    username?: string;
-    credential?: string;
-}
-
-export interface WHIPResourceSsrc {
-    ssrc: string;
-    cname?: string;
-    mslabel?: string;
-    label?: string;
-}
-
-export interface WHIPResourceSsrcGroup {
-    semantics: string;
-    ssrcs: string[];
-}
-
-export interface WHIPResourceMediaStreams {
-    audio: {
-        ssrcs: WHIPResourceSsrc[];
-    },
-    video: {
-        ssrcs: WHIPResourceSsrc[];
-        ssrcGroups: WHIPResourceSsrcGroup[];
-    }
-}
-
-export interface WHIPResource {
-    connect();
-    getProtocolExtensions(): string[];
-    sdpAnswer(): Promise<string>;
-    assignBroadcaster(broadcaster: Broadcaster);
-    getIceServers(): WHIPResourceICEServer[];
-    getId(): string;
-    getETag(): string | undefined;
-    getType(): string;
-    patch(body: string, eTag?: string): Promise<number>;
-    getMediaStreams(): WHIPResourceMediaStreams;
-    destroy(): void;
-}
