+++ packages/server/src/whip/wrtc/broadcasterWrtcWhipResource.ts
@@ -1,15 +1,15 @@
 import { MediaStream } from "wrtc";
-import { WHIPResourceICEServer, IANA_PREFIX } from "../models/WHIPResource";
-import { WRTCWHIPResource } from "./WRTCWHIPResource";
+import { WhipResourceIceServer, IANA_PREFIX } from "../whipResource";
+import { WrtcWhipResource } from "./wrtcWhipResource";
 
-interface WRTCBroadcasterOptions {
+interface BroadcasterWrtcWhipResourceOptions {
   channelId?: string;
 }
 
-export class WRTCBroadcaster extends WRTCWHIPResource {
+export class BroadcasterWrtcWhipResource extends WrtcWhipResource {
   private channelId: string;
 
-  constructor(sdpOffer: string, iceServers?: WHIPResourceICEServer[], opts?: WRTCBroadcasterOptions) {
+  constructor(sdpOffer: string, iceServers?: WhipResourceIceServer[], opts?: BroadcasterWrtcWhipResourceOptions) {
     super(sdpOffer, iceServers);
     this.channelId = this.getId();
     if (opts?.channelId) {
