+++ packages/server/src/whip/wrtc/rtspWrtcWhipResource.ts
@@ -1,6 +1,6 @@
-import { WRTCWHIPResource } from "./WRTCWHIPResource";
-import { MPEGTS, MPEGTSResolution } from "../transform/mpegts";
-import { WHIPResourceICEServer, IANA_PREFIX } from "../models/WHIPResource";
+import { WrtcWhipResource } from "./wrtcWhipResource";
+import { MPEGTS, MPEGTSResolution } from "../../transform/mpegts";
+import { WhipResourceIceServer, IANA_PREFIX } from "../whipResource";
 
 import ffmpeg from "fluent-ffmpeg";
 
@@ -26,17 +26,17 @@ export class RTSPResolution {
   }
 }
 
-interface WRTCRTSPOptions {
+interface RtspWrtcWhipResourceOptions {
   server: string;
   resolution: RTSPResolution;
 }
 
-export class WRTCRTSP extends WRTCWHIPResource {
+export class RtspWrtcWhipResource extends WrtcWhipResource {
   private rtspServer: string;
   private output;
   private outputResolution: RTSPResolution;
 
-  constructor(sdpOffer: string, iceServers?: WHIPResourceICEServer[], opts?: WRTCRTSPOptions) {
+  constructor(sdpOffer: string, iceServers?: WhipResourceIceServer[], opts?: RtspWrtcWhipResourceOptions) {
     super(sdpOffer, iceServers);
     this.rtspServer = "rtsp://127.0.0.1:8554";
     this.outputResolution = new RTSPResolution(960, 540);
