+++ packages/server/src/whpp/whppRequests.ts
@@ -1,0 +1,16 @@
+export interface WhppAnswerRequest {
+  answer: string;
+}
+
+export interface WhppCandidateRequest {
+  candidate: string;
+}
+
+export interface WhppMediaStream {
+  streamId: string;
+}
+
+export interface WhppOfferResponse {
+  offer: string;
+  mediaStreams: WhppMediaStream[];
+}
