+++ packages/server/src/whpp/wrtc/wrtcWhppViewer.ts
@@ -1,20 +1,20 @@
 import { RTCPeerConnection } from "wrtc";
 import { v4 as uuidv4 } from "uuid";
 
-import { BroadcasterICEServer } from ".";
+import { BroadcasterIceServer } from "../../broadcaster";
 import { EventEmitter } from "events";
-import { ViewerAnswerRequest, ViewerCandidateRequest, ViewerMediaStream, ViewerOfferResponse } from './ViewerRequests'
+import { WhppAnswerRequest, WhppCandidateRequest, WhppMediaStream, WhppOfferResponse } from '../whppRequests'
 import { parse } from 'sdp-transform'
-import { Viewer } from './Viewer'
+import { WhppViewer } from '../whppViewer'
 
 const ICE_GATHERING_TIMEOUT = process.env.ICE_GATHERING_TIMEOUT ? parseInt(process.env.ICE_GATHERING_TIMEOUT) : 4000;
 const CONNECTION_TIMEOUT = 60 * 1000;
 
 export interface ViewerOptions {
-  iceServers?: BroadcasterICEServer[];
+  iceServers?: BroadcasterIceServer[];
 }
 
-export class WRTCViewer extends EventEmitter implements Viewer {
+export class WrtcWhppViewer extends EventEmitter implements WhppViewer {
   private channelId: string;
   private viewerId: string;
   private peer: RTCPeerConnection;
@@ -104,7 +104,7 @@ export class WRTCViewer extends EventEmitter implements Viewer {
     return this.viewerId;
   }
 
-  async handlePost(stream: MediaStream): Promise<ViewerOfferResponse> {
+  async handlePost(stream: MediaStream): Promise<WhppOfferResponse> {
     for (const track of stream.getTracks()) {
       this.log(`Added local track ${track.kind} from ${this.channelId}`);
       this.peer.addTrack(track, stream);
@@ -114,7 +114,7 @@ export class WRTCViewer extends EventEmitter implements Viewer {
     await this.peer.setLocalDescription(offer);
     await this.waitUntilIceGatheringStateComplete();
 
-    const viewerResponse: ViewerOfferResponse = {
+    const viewerResponse: WhppOfferResponse = {
       offer: this.peer.localDescription.sdp,
       mediaStreams: this.getMediaStreams()
     };
@@ -129,21 +129,21 @@ export class WRTCViewer extends EventEmitter implements Viewer {
     return viewerResponse;
   }
 
-  async handlePut(request: ViewerAnswerRequest): Promise<void> {
+  async handlePut(request: WhppAnswerRequest): Promise<void> {
     await this.peer.setRemoteDescription({
       type: 'answer',
       sdp: request.answer
     });
   }
 
-  async handlePatch(request: ViewerCandidateRequest): Promise<void> {
+  async handlePatch(request: WhppCandidateRequest): Promise<void> {
     await this.peer.addIceCandidate({ candidate: request.candidate });
   }
 
-  private getMediaStreams(): ViewerMediaStream[] {
+  private getMediaStreams(): WhppMediaStream[] {
     const parsedSdp = parse(this.peer.localDescription.sdp);
     let storedStreamIds = new Set<string>();
-    let mediaStreams: ViewerMediaStream[] = [];
+    let mediaStreams: WhppMediaStream[] = [];
 
     for (let media of parsedSdp.media) {
       const msidSplit = media.msid && media.msid.split(' ');
